# guest: Add proto.Metadata and fix prefilter unmarshalling