--- a/internal/e2e/go.mod
+++ b/internal/e2e/go.mod
@@ -6,7 +6,7 @@ go 1.20
 require (
 	github.com/labstack/echo v3.3.10+incompatible
 	github.com/stealthrocket/wzprof v0.1.5
-	github.com/tetratelabs/wazero v1.3.0
+	github.com/tetratelabs/wazero v1.3.1
 	go.uber.org/zap v1.19.0
 	k8s.io/api v0.27.3
 	k8s.io/apimachinery v0.27.3