… examples/nodenumber/main.wasm
… guest/testdata/all/main.wasm
… guest/testdata/cyclestate/main.wasm
… guest/testdata/filter/main.wasm
… guest/testdata/score/main.wasm
… internal/e2e/go.sum
… scheduler/go.sum