#### What type of PR is this?

/kind cleanup

#### What this PR does / why we need it:

This adds `proto.Metadata` type which can later be used for logging, metrics or stringer functions. This also fixes a prefilter unmarshal bug where we used the wrong host function to decode the data.

This also updates wazero patch, which isn't related, but saves a PR.

#### Which issue(s) this PR fixes:

NONE

#### Special notes for your reviewer:

#### Does this PR introduce a user-facing change?

NONE

#### What are the benchmark results of this change?

```benchstat
goos: darwin
goarch: arm64
pkg: sigs.k8s.io/kube-scheduler-wasm-extension/internal/e2e/scheduler
                          │ before.txt  │              now.txt               │
                          │   sec/op    │    sec/op     vs base              │
Example_NodeNumber/New-12   74.18m ± 1%   73.66m ±  1%       ~ (p=0.180 n=6)
Example_NodeNumber/Run-12   46.28µ ± 1%   46.36µ ± 10%       ~ (p=0.394 n=6)
geomean                     1.853m        1.848m        -0.26%

```